pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitable-agent-fabric"
version = "0.1.0"
description = "Scene-adaptive multi-agent fabric on top of Feishu Bitable"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
baf = "baf.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
